requirements.txt
absl-py==1.1.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
boto3==1.34.69
botocore==1.34.69
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cycler==0.10.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
fire==0.6.0
Flask==3.0.2
flatbuffers==2.0
fonttools==4.34.2
fsspec==2024.3.1
gast==0.5.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.47.0
h5py==3.6.0
huggingface-hub==0.21.4
idna==2.10
imbalanced-learn==0.9.1
imblearn==0.0
imutils==0.5.4
ipykernel==6.17.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jason==0.1.7
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-http-over-ws==0.0.8
jupyter-server==1.23.2
jupyter_client==7.4.6
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface @ git+https://github.com/rcmalli/keras-vggface.git@719ad75625fe51436b9a9479114b3d136847e11a
kiwisolver==1.4.3
libclang==13.0.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==2.0.4
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycparser==2.21
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2022.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.25
rsa==4.8
s3transfer==0.10.1
safetensors==0.4.2
sahi==0.11.15
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
shapely==2.0.3
six==1.16.0
sklearn==0.0.post1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.3.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0rc1
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.17.0
terminaltables==3.1.10
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.2
torch==1.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.39.1
typing_extensions==4.3.0
ultralytics==8.1.34
urllib3==1.26.9
vboxapi==1.0
virtualenv==20.25.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.1
widgetsnbextension==4.0.3
wrapt==1.13.3
yolov5==7.0.13
